Step one of setting media to be authority on call data.
call.hrl has queued_call added as a record; queued_call being
the data a call queue uses.

call_queue.hrl uses a stripped down call record, queued_call.  To
queue a call, on passes in a pid of a media process, which can
return a proper call record.

dummy_media is meant to act as a media process.

util has the magic skill '_all' added to the list of magic skills.

All other modules listed were modified to account for adding a call
pid to a call_queue instead of a call record.