Updated multinode tests to account for cook changes.
The cook now checks to see if the media pid passed is alive.  However,
you can only use is_process_alive on a pid that is on the same local
node.  Thus, if media is created and queued on a different node, the cook
will fail to start with a badarg error.