freeswitch media is now in charge of ringouts
To make ringouts more accurate (both in time and outband ring types),
media is now in charge of stating when a ring-out happens.
This means that when a media is told to start ringing, the ringout time
is passed in (in milliseconds).
This alsmo means the media needs to extract the cook, and tell the cook
when to stop ringing an agent.