Enhancing configurability
cpx_supervisor now accapts a list of nodes as it's start paramter.
This is passed into agent_manager and queue_manager as is to get around
gen_leader weaknesses.

cpx.erl looks up the nodes list from an applicaiton eviroment file, and
passes that into the start functions of cpx_supervisor.