Expanding base for ring channel support
freeswitch_ring takes the place of the recursive fun used earlier.
It takes the messages from the call created by the originate call
in freeswitch_media, and is in charge of set the agent state to oncall,
as well as ensuring the agent state is set correctly when the call ends.