Make magic skills expand to {Atom, Value} instead of just Value
For example _node now expands to {'_node', node@host} instead of just
node@host. This makes it a lot simpler to add/remove skills to an
already expanded list without creating duplicates. It also lets us
ensure that only one instance of each magic skill is present in a
skill list (so that a call couldn't be tagged for two brands at once.

Also added .swo files to .gitignore