Health determination algorythm modified.
Health is not to be passed in as a single number, but an object who's
keys are points of interest, and the values being the established 0 to 100
scale.  This is used in conjunction with the 'children' of an item of
interest.  eg:  nodes take from agent profiles and queue groups.  queue
groups take from the queues, and so forth.