Fixed mail_map and death handling.
Mail_map was often missing the required timestamp field.  Also, now
the media will be able to find the manager if it dies.  In addition, the
manager will die if it can't find nor start the smtp server.  To acheive
this, the server needed to get a name registered to it.  Killing the server
kills the manager, but killing the manage does not kill the server, thus the
need for a registered name to look for.