Find_untermed and minor refactor.
the find_untermed function is part of the overall rewrite to get
cdr's to be more sane and complete.

the minor refactor is that "catch gen_event:notify" was in almost every
event function.  So I factored that out and put it in it's own function,
event/1.  This will allow me to make generic caught calls as well.