restructed agent_auth intgration hooks.
agent_auth is no longer a gen_server, as it doesn't need to be.
It makes user the integration module for querying the integration
server that may be running.  Thus, the build tables call has been moved
to agent_manager.  Tests redone to reflect the change in functionality.
Has same behaviour as before, in that if integration fails, it falls back
to local.

passwords are mostly handled in plain test; it is expected that endpoints
will appropriately secure them before sending them over the wire.