Corrected email_media startups procedure.
At least where putting mail from a file into queue were concerned.
Now, the onus of checking if an email is mime mail is on the email_media
that is trying to start up, rather than the email_media_manager.
ie:  now if the mimemail:decode fails, a process we don't care about
(email media) crashes and burns, instead of the manager.