Dummy media managers triumphant return!
Not longer in the role gen_server_mock filled, it's been
wandering the wastelands, but now it's back, and hungry for
vengence!

Actually it's to act as a manager for dummy_media to better simulate
how other medias and thier managers behave.  The idea is that
the media managers will monitor thier medias, and if one dies inform
cpx_mon of when the funeral will be.