Added {voicemail, State} return to gen_media
If the callback returns {voicemail, State} from a general handle_call,
handle_cast, or handle_info, and the media is in a
queue, gen_media will remove the media from queue and stop it from
ringing to any agents.  All 3 returns go through the same code, including
the media -> callback version of voicemail.