Prevent spurious logoffs due to "missed polls"
The reason that agents were occasionally incorrectly logged off is
becuase when an event came in close to the time when a poll was
expiring, it'd send it out, clear the poll pid but not update the poll
pid time. This meant that when the periodic timeout check ran, the last
time it saw was when the poll pid was *established* which could well be
longer ago than the cutoff, thus triggering a logout.