Megapatch for universal caseid support
This allows all currently implemented medias to support being tagged
with a CRM Case ID. The Case ID is currently requested when an agent
tries to transfer a media to a queue or another agent so that the other
agent can easily pull up the information on what has already been done.

This also changes the behaviour of agent transfers in that now a URL pop
is done in the same way as if the call was being delivered to them from
queue.