Strategy to deal w/ busy systems handling many events.
When an event enters the poll_queue on the agent_web_connection, a 500ms
countdown is started.  Further events do not reset this count.  Once the
count reaches 0, the poll_queue is flushed.  This means the queue gets
flushed eveny 500 ms regularly on a very busy system.  The counter side is
this introduces a 500 ms delay into any ui actions that relay on the poll
queue.