Agent list rotation implemented.
This is to github issue 5:  Agent ring selection conflicting.
each time the agent manager gets a request for the agent list for routing
and the agent list has not changed, an internal counter is incremented.
That count is reset each time an agent changes availability or skills.  The
cook then uses this count to rotate the list, so if a ton of calls come in,
offers are distributed evenly amoung all elibible agents rather than just
The first agent.  Each node's list is rotated independantly of each each
other.