Added url pop on agent pick-up.  Needs attn.
The urls are currently hard-coded in the macro URL_POPS.  Urls_pops is a
proplist of {WindowName, Url}.  The actual call to url_pop can take a
proplist of arguments for string interplation of the url, but there was no
idication of what variables should go there.  It's not difficult to address
however.