Trying to make a dispatcher more in charge of its death
Monitor the cook, 'cause when that dies, the call isn't in queue
any more, so the dispatchers purpose is out.  Balance can now only
launch up dispatchers, so 'zombie' dispatchers eventually get reused.  As
calls complete, and the number of agents dwindle, the number of dispatchers
will dwindle as well.  One will die, but not be respawned on balance.
As for the cook, no need for an extra message when the dispatcher can
tell itself to regrab.