Allowing for test:profile
Profiling tests tend to take a long time, and generally don't need to
be run that often.  Those tests can be placed in an ifdef(PROFILE).  To get
this to work, however, regular tests need to be placed under something
other than TEST as eunit will always define that if included.  Eunit is
used for it's fixture capability, so it's inadvisable just to turn it off.
So, if TEST is defined, define a second macro at that point, and place
usual tests behind that.

Of course, this may all be for naught if some better solution comes around.