A dev replacement for boot.rb.
rebar generate does not like to generate if it's already be generated,
making the rebar created bin less than idea, espcially considering the
libs don't get moved in that case.  boot.rb is on the way out to remove the
dependance on ruby.  So for just a quick and dirty start that will play
nicely with rebar compile and l/1, this will do.