Commit 'avoid queue starvation' patch
This adds a 'last service time' to the queue which we can use to avoid
starving a queue at the expense of another at the same weight. Consider
the following example:

You have 2 queues, A and B. A and B have no common agents and have the
same weight. A gets a *lot* of calls and B is very infrequently hit.
Without tracking service level, A is often going to win out over B
because its got more calls that have a longer hold time, even if B has
agents doing nothing but waiting for calls and A is out of agents, we'll
keep trying to offer calls out of A because of the bigger hold time of
the first call. Calls in B won't be offered until the call in B gets
older than the oldest call in A (there's a random bias here, but its not
enough).

The addition of a service time helps avoid starvation like this by
sorting on the difference between age of the oldest call and the last
service time. So if there's an undeliverable call in a queue, since the
age of the call will be high, it won't be offered constantly because the
service time will be recent.