Made the plugin system a bit saner and more otp-y.
Based on the model used by RabbitMQ.  Added load, reload, unload, and
running functions so plugins can be managed a bit.  OpenACD still doesn't
supervise the apps, but it does nicely start them.  It keeps a list of the
top level apps, but does start the required apps for each plugin (kinda
ignoring them).

This _will_break the existing sipx plugin, but the solution is simple:
move the sipx deps into the same level as the sipx directory.  Install them
all together.