Fixed code flow informing agent_manager when agent available.
If an agent goes into ring_lock, it no longer tells the agent_manager
it is avialable until ring_lock is cleared.  This prevents the agent_manager
from presenting the agent as a valid ring target, causing a ring to fail.

Primarily is helps the edge case of one agent and one call; the agent could
end up in ring_lock, refuse to go ringing, and then the call would never
ring to that agent again.  Now, the call only rings to the agent when the
agent can actually answer it.