Fixing race condition on agent_transfer
If the command to bridge the call comes in too quickly after the call
is answered, freeswitch will kill the caller leg saying there's an error
doing an RTP reinvite.  The best solution at the moment is to put in a
pause.  If/when a better solution presents itself, hopefully this will be
corrected.